Address Micah's PR comments
